moved pretty_xml from repository.supermodel to xml_utils
