+++ metashare/repository/editor/resource_editor.py
@@ -233,7 +233,7 @@ def export_xml_resources(modeladmin, request, queryset):
     from StringIO import StringIO
     from zipfile import ZipFile
     from xml.etree import ElementTree
-    from metashare.repository.supermodel import pretty_xml
+    from metashare.xml_utils import pretty_xml
     from django import http
 
     zipfilename = "resources_export.zip"
@@ -458,7 +458,7 @@ class ResourceModelAdmin(SchemaModelAdmin):
               % {'name': force_unicode(opts.verbose_name), 'key': escape(object_id)})
 
         from xml.etree import ElementTree
-        from metashare.repository.supermodel import pretty_xml
+        from metashare.xml_utils import pretty_xml
         from django import http
 
         try:
