+++ metashare/repository/models.py
@@ -9,7 +9,7 @@ from xml.etree.ElementTree import tostring
 
 # pylint: disable-msg=W0611
 from metashare.repository.supermodel import SchemaModel, SubclassableModel, \
-  _make_choices_from_list, InvisibleStringModel, pretty_xml, \
+  _make_choices_from_list, InvisibleStringModel, \
   REQUIRED, OPTIONAL, RECOMMENDED
 from metashare.repository.editor.widgets import MultiFieldWidget
 from metashare.repository.fields import MultiTextField, MetaBooleanField, \
@@ -19,6 +19,7 @@ from metashare.repository.validators import validate_lang_code_keys
 from metashare.storage.models import StorageObject
 
 from metashare.settings import DJANGO_BASE, LOG_LEVEL, LOG_HANDLER
+from metashare.xml_utils import pretty_xml
 
 # Setup logging support.
 logging.basicConfig(level=LOG_LEVEL)
